working ingest service
