+++ examples/quick-start/packages/ingest/src/ingest.config.ts
@@ -4,63 +4,19 @@ import {
   makeMongoDbEmbeddedContentStore,
   makeMongoDbPageStore,
 } from "mongodb-rag-core";
-import { strict as assert } from "assert";
 import { standardChunkFrontMatterUpdater } from "mongodb-rag-ingest/embed";
-import {
-  MakeMdOnGithubDataSourceParams,
-  makeMdOnGithubDataSource,
-} from "mongodb-rag-ingest/sources";
-import dotenv from "dotenv";
 import path from "path";
+import { loadEnvVars } from "./loadEnvVars";
+import { mongoDbChatbotFrameworkDocsDataSourceConstructor } from "./mongodbChatbotFrameworkDataSource";
 
-// Load environment variables from .env file
-dotenv.config({ path: path.join(__dirname, "..", "..", "..", ".env") });
+// Load project environment variables
+const dotenvPath = path.join(__dirname, "..", "..", "..", ".env"); // .env at project root
 const {
   MONGODB_CONNECTION_URI,
   MONGODB_DATABASE_NAME,
   OPENAI_API_KEY,
-  OPENAI_EMBEDDING_DEPLOYMENT,
-} = process.env;
-assert(MONGODB_CONNECTION_URI, "MONGODB_CONNECTION_URI is required");
-assert(MONGODB_DATABASE_NAME, "MONGODB_DATABASE_NAME is required");
-assert(OPENAI_API_KEY, "OPENAI_API_KEY is required");
-assert(OPENAI_EMBEDDING_DEPLOYMENT, "OPENAI_EMBEDDING_DEPLOYMENT is required");
-
-// Data source to get the docs content.
-export const mongodbRagFrameworkDocsConfig: MakeMdOnGithubDataSourceParams = {
-  name: "mongodb-rag-framework",
-  repoUrl: "https://github.com/mongodb/chatbot/",
-  repoLoaderOptions: {
-    branch: "main",
-  },
-  filter: (path) =>
-    path.startsWith("docs/") &&
-    path.endsWith(".md") &&
-    !path.endsWith("README.md"),
-  pathToPageUrl(pathInRepo) {
-    const baseUrl = "https://mongodb.github.io/chatbot/";
-    const path = pathInRepo.replace(/^docs\/docs\//, "").replace(/\.md$/, "");
-    return `${baseUrl}${path}`;
-  },
-  extractTitle: (pageContent, frontmatter) =>
-    (frontmatter?.title as string) ?? extractFirstH1(pageContent),
-};
-
-// Helper function
-function extractFirstH1(markdownText: string) {
-  const lines = markdownText.split("\n");
-
-  for (let line of lines) {
-    if (line.startsWith("# ")) {
-      // Remove '# ' and any leading or trailing whitespace
-      return line.substring(2).trim();
-    }
-  }
-  return null;
-}
-const mongodbRagFrameworkDocsSourceConstructor = () => {
-  return makeMdOnGithubDataSource(mongodbRagFrameworkDocsConfig);
-};
+  OPENAI_EMBEDDING_MODEL,
+} = loadEnvVars(dotenvPath);
 
 export default {
   embedder: async () => {
@@ -69,7 +25,7 @@ export default {
     const { OpenAIClient, OpenAIKeyCredential } = await import("@azure/openai");
     return makeOpenAiEmbedder({
       openAiClient: new OpenAIClient(new OpenAIKeyCredential(OPENAI_API_KEY)),
-      deployment: OPENAI_EMBEDDING_DEPLOYMENT,
+      deployment: OPENAI_EMBEDDING_MODEL,
       backoffOptions: {
         numOfAttempts: 25,
         startingDelay: 1000,
@@ -97,9 +53,9 @@ export default {
   }),
   // Add data sources here
   dataSources: async () => {
-    const mongodbRagFrameworkSource =
-      await mongodbRagFrameworkDocsSourceConstructor();
+    const mongodbChatbotFrameworkSource =
+      await mongoDbChatbotFrameworkDocsDataSourceConstructor();
 
-    return [mongodbRagFrameworkSource];
+    return [mongodbChatbotFrameworkSource];
   },
 } satisfies Config;
